+++ instrumentation/apache-httpclient-4.0/src/main/java/io/opentelemetry/instrumentation/hypertrace/apache/httpclient/ApacheHttpClientObjectRegistry.java
@@ -1,0 +1,12 @@
+package io.opentelemetry.instrumentation.hypertrace.apache.httpclient;
+
+import io.opentelemetry.api.trace.Span;
+import java.util.WeakHashMap;
+
+public class ApacheHttpClientObjectRegistry {
+
+  private ApacheHttpClientObjectRegistry() {}
+
+  public static final WeakHashMap<Object, Span> objectToSpanMap = new WeakHashMap<>();
+
+}
